# src/main/java/com/radian/cuwb/rest/api/UserApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.18).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.radian.cuwb.rest.api;

import com.radian.cuwb.rest.model.Response;
import java.util.UUID;
import com.radian.cuwb.rest.model.User;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2022-06-08T13:19:07.446+04:00[Asia/Muscat]")
@Api(value = "user", description = "the user API")
public interface UserApi {

    @ApiOperation(value = "Get User", nickname = "getUser", notes = "Get User", response = User.class, tags={ "user", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Succesfuly Created", response = User.class),
        @ApiResponse(code = 400, message = "Invalid  Data", response = Response.class) })
    @RequestMapping(value = "/user",
        produces = { "application/json" }, 
        consumes = { "*/*" },
        method = RequestMethod.GET)
    ResponseEntity<User> getUser(@ApiParam(value = "Auth object" ,required=true )  @Valid @RequestBody User body
,@ApiParam(value = "Request Id" ,required=true) @RequestHeader(value="requestId", required=true) UUID requestId
,@ApiParam(value = "Channel Name" ,required=true) @RequestHeader(value="channel", required=true) String channel
,@ApiParam(value = "Time date Unix Format" ,required=true) @RequestHeader(value="timeStamp", required=true) Integer timeStamp
,@ApiParam(value = "Autorization Token" ) @RequestHeader(value="authToken", required=false) String authToken
,@ApiParam(value = "Version of API" ) @RequestHeader(value="version", required=false) String version
,@ApiParam(value = "Entity" ) @RequestHeader(value="entity", required=false) String entity
);

}